+++ app/models/idea.rb
@@ -1,0 +1,3 @@
+class Idea < ApplicationRecord
+    has_many :goals
+end