+++ config/initializers/idea_icons.rb
@@ -1,0 +1,1 @@
+IDEAS = YAML.load_file(Rails.root.join('config', 'idea_icons.yml')).with_indifferent_access