+++ db/migrate/20250806224051_create_ideas.rb
@@ -1,0 +1,9 @@
+class CreateIdeas < ActiveRecord::Migration[7.1]
+  def change
+    create_table :ideas do |t|
+      t.string :title
+
+      t.timestamps
+    end
+  end
+end