+++ db/migrate/20250806224657_add_idea_to_goals.rb
@@ -1,0 +1,5 @@
+class AddIdeaToGoals < ActiveRecord::Migration[7.1]
+  def change
+    add_reference :goals, :idea, null: false, foreign_key: true
+  end
+end